Break out classes and interfaces into own files
