+++ ts/CobrowseAccessibilityService.d.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-extraneous-class,accessor-pairs */
+export default class CobrowseAccessibilityService {
+  static showSetup (): Promise<void>
+
+  static isRunning (): Promise<boolean>
+}
